downloadImage for ImageView externalization
