+++ app/src/main/java/com/sushi/izishopping/DownloadImageFromUrl.kt
@@ -1,0 +1,30 @@
+package com.sushi.izishopping
+
+import android.annotation.SuppressLint
+import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.os.AsyncTask
+import android.util.Log
+import android.widget.ImageView
+import java.net.URL
+
+class DownloadImageFromUrl(@field:SuppressLint("StaticFieldLeak") var bmImage: ImageView) :
+    AsyncTask<String?, Void?, Bitmap?>() {
+    override fun doInBackground(vararg params: String?): Bitmap? {
+        val urldisplay = params[0]
+        var mIcon11: Bitmap? = null
+        try {
+            val `in` = URL(urldisplay).openStream()
+            mIcon11 = BitmapFactory.decodeStream(`in`)
+        } catch (e: Exception) {
+            Log.e("Error", e.message!!)
+            e.printStackTrace()
+        }
+        return mIcon11
+    }
+
+    override fun onPostExecute(result: Bitmap?) {
+        bmImage.setImageBitmap(result)
+    }
+
+}
