+++ app/src/main/java/com/sushi/izishopping/foodlist/FoodAdapter.kt
@@ -1,18 +1,13 @@
 package com.sushi.izishopping.foodlist
 
-import android.annotation.SuppressLint
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.os.AsyncTask
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import com.sushi.izishopping.Food
 import com.sushi.izishopping.databinding.ItemFoodBinding
+import com.sushi.izishopping.DownloadImageFromUrl
 import java.io.*
-import java.net.URL
+
 
 
 // prend en parametre du constructeur, la FoodList
@@ -34,7 +29,7 @@ class FoodAdapter(private var foodList: List<Food>)
             foodTitleTextView.text = food.name
             scanDateTextView.text = "TBD"
             barCodeTextView.text = food.barcode
-            DownloadImageTask(foodImageView).execute(food.imgLink)
+            DownloadImageFromUrl(foodImageView).execute(food.imgLink)
         }
     }
 
@@ -44,25 +39,4 @@ class FoodAdapter(private var foodList: List<Food>)
         this.foodList = foodList
         notifyDataSetChanged()
     }
-}
-
-private open class DownloadImageTask(@field:SuppressLint("StaticFieldLeak") var bmImage: ImageView) :
-    AsyncTask<String?, Void?, Bitmap?>() {
-    override fun doInBackground(vararg params: String?): Bitmap? {
-        val urldisplay = params[0]
-        var mIcon11: Bitmap? = null
-        try {
-            val `in` = URL(urldisplay).openStream()
-            mIcon11 = BitmapFactory.decodeStream(`in`)
-        } catch (e: Exception) {
-            Log.e("Error", e.message!!)
-            e.printStackTrace()
-        }
-        return mIcon11
-    }
-
-    override fun onPostExecute(result: Bitmap?) {
-        bmImage.setImageBitmap(result)
-    }
-
 }
